Fix bug that was leading to label line number count being off
Prior to this commit comments and blanklines were being counted
in the function that numbered lines. This would throw off the jump
commands because they would point the the wrong location in the
program. Also updates how lines are counted and skips incrementing
the line count when a label is seen. This prevents an issue where
counting labels as final program lines would cause the jump commands
to point to the wrong place.