Adjust how Static segments are addressed
Each .vm file needs access to a seperate Static memory segment. This commit
changes Static memory handling to use <fileName>$<offest> naming for
each static segment that is addressed in a file. Prior to this commit
the Static segments were addressed directly using the base address of 16
and adding the offset this caused the segments to shadow each other when
there were mulitple .vm files.