Fix literal word matching helper
Prior to this commit matches for certain literals could fail if
the word they were attempting to match had a similar prefix. This
is due to Parsec.string consuming characters from the incoming
string when they match. Parsec.choice chooses the first rule
which starts to consume characters. This could cause a keyword like
'constructor' to match the rule for 'class' because they both start
with 'c', resulting in an unintended parse error.

This commit uses Parsec.try to wrap the string parsers, this attempts
a parse but doesn't consume any characters from the incoming string on
a failure. Parsec.choice can then feed the string after a partial failure
to the next parser without issue.