Add parser support for Jack terms
This commit adjusts the types the parser produces. It moves terminal
types which are only seen in Jack 'terms' under the Term type. This is
needed so the types line up when combining the individual parsers for
each terminal element. It also includes a type and parser for expressions
which are included in Jack 'terms'.