Fix overly permissive rule in identifier parser
Prior to this commit the identifier parser could allow a non-letter
char in the first position. For example it would parse '(foo' because
the rule for the first char only specified that it couldn't be a digit.
This caused the higher level termP parser to misclassify a ')' as a
VarName. This commit changes the rule so only letters and '_' are
allowed in the first position.