Reorder term parser choices
Prior to this commit the ordering of parsers in termP was incorrect.
This would result in an input like "foo.bar()" to be partially consumed
resulting in a (Varname "foo") with ".bar()" remaining in the input
string.

This commit changes the order of the parsers in termP so that the
more selective parsers are tried first.