Add test helper to avoid needing default type in fromRight
Prior to this commit fromRight was used to extract the parsed values
returned from Parsec. This required each test to have a default value
of a matching type. Constructing these defaults became a pain as the
size of parsed results grew. The added helper extracts the Right value
and throws it in a Maybe to avoid this issue.