Use binary tree to model Jack expressions
This commit changes the Expr data type to use a binary tree to model
Jack expressions. Prior to this commit an Expr was a Term followed
by a list of (Op, Term)s. This strucutre was fine for the XML generation
but a binary tree will be more natrual during code generation. A depth
first traversal of the Expr tree will return the expression elements in
the order that's needed when pushing these elements on the stack.