Add support for VM code gen of Jack expressions
This commit creates a VMGen typeclass which defines a genVM interface.
This interface will be used by all the types in AST to generate VM commands.
The commit also handles the post order traveral of Term trees. This is
nessesary to produce VM commands which manipulate the stack to compute
the higher level Jack expressions.