Rewrite genVM for Terms using applicative functors
This commit changes how the genVM calls for Terms concat values from
their sub elements. Before this commit the do notation was used and
each computed value for a Term element was unwrapped out of the
SymbolTableState contex, concatenated as Strings and then rewrapped
in the SymbolTableState context.

This commit takes advantage of the fact that the State Monad is also
an Applicative Functor. This allows the fmaping of the (++) function
using <$> resulting in a partial funtion that's wrapped in the context,
this paritial function is then given the last wrapped argument via <*>.
The result of doing this means we no longer need to unwrap the string
values out of the SymbolTableState context in order to call functions
on them.