Add support for code generation of Do statements
This commit moves the handling of SubroutineCalls into its own
VMGen instance so it can be shared by data types outside of Term
and adds support for Do statememts.

This commit also fixes a bug where identifiers like className.methodName
were being looked up in the symbol table. There are never symbol table
entries created for class or method names so this caused each lookup to
produce an "error not found" string in the generated VM commands. This
no longer happens as class and method names are now handled driectly and
bypass the symbol table lookup.