Use sequence to simplify code generation for compound types
This commit uses sequence to flip a list of monadic values into a monad
containing a list of those values. In this case it transforms
[SymbolTableState String] -> SymbolTableState [String] then concat is used
to combine the strings inside the monad producing one String of VM commands.
Doing this greatly simplified the handling of compound types and allowed
most of the generation to be done without needing to unwrap/rewrap values
inside of a do block.