Throw error when symbol not found in symbol tables
This commit causes an error to be thrown during code generation if a
symbol isn't found in the local or class level symbol tables. In this
case crashing the program is the behavior we want because there isn't
a sensible thing to do if the info for a symbol doesn't exist and we
want to alert a user before producing incorrect VM code.