Handle class name lookups during code generation for method calls
This commit changes the behavior of code generation when a method is
called on an object. See example below:

var ClassFoo c1;
let c1 = ClassFoo.new();
do c1.FooMethod;

In this example code should be generated that pushes the base address
of the c1 object before calling the ClassFoo.FooMethod function.
This commit uses the symbol tables to find an entry for c1 and then uses the
type of that entry to find the class where the function is defined. If the
class part of a <className>.<methodName> call is not found in a symbol table
the call is left unchanged and nothing extra is pushed on to the stack.