Update handling of array access in statements
This commit updates the handling of array access when the value at the
array index is needed. Example:
let foo = a[1];
Here we need to access the value found at index 1 of the 'a' array. The
code generated for the right side of the expression above uses Pointer 1
and 'that' to access this value and pushes it on the stack.