Move handling of return code gen to Statement level
This commit moves the code generation caused by any return statement which
returns a value to the Statement data type. Previously the 'return' VM command
was being generated when handling the higher level SubroutineDec data type
because of special handling needed in void methods. This was incorrect because
there can be multiple return statements in a function and they need to produce
'return' VM calls independently. For example:

if (~foo) {
  return bar;
} else {
  return baz;
  }

return statements without values are now assumed to be returns from void
methods and handled by pushing 0 onto the stack and then calling return.